requirements.txt
pylint==3.0.2
pytest==7.4.3
python-decouple==3.8
fastapi[standard]==0.112.2
SQLAlchemy==2.0.23
psycopg[binary,pool]==3.1.14
psycopg2-binary==2.9.9
redis==5.0.4
pydantic==2.8
plutonkit==1.0.0a22
pyjwt==2.9.0
httpx==0.27.0
isort
